+++ flight/targets/FlyingF3/System/pios_board.c
@@ -141,7 +141,7 @@ static const struct pios_lsm303_cfg pios_lsm303_cfg = {
  *  eg. PWM, PPM, GCS, SPEKTRUM1, SPEKTRUM2, SBUS
  * NOTE: No slot in this map for NONE.
  */
-uint32_t pios_rcvr_group_map[MANUALCONTROLSETTINGS_CHANNELGROUPS_NONE];
+uintptr_t pios_rcvr_group_map[MANUALCONTROLSETTINGS_CHANNELGROUPS_NONE];
 
 #define PIOS_COM_TELEM_RF_RX_BUF_LEN 512
 #define PIOS_COM_TELEM_RF_TX_BUF_LEN 512
@@ -186,7 +186,7 @@ uintptr_t pios_can_id;
 static void PIOS_Board_configure_com (const struct pios_usart_cfg *usart_port_cfg, size_t rx_buf_len, size_t tx_buf_len,
 		const struct pios_com_driver *com_driver, uintptr_t *pios_com_id)
 {
-	uint32_t pios_usart_id;
+	uintptr_t pios_usart_id;
 	if (PIOS_USART_Init(&pios_usart_id, usart_port_cfg)) {
 		PIOS_Assert(0);
 	}
@@ -220,18 +220,18 @@ static void PIOS_Board_configure_dsm(const struct pios_usart_cfg *pios_usart_dsm
 		const struct pios_com_driver *pios_usart_com_driver,enum pios_dsm_proto *proto,
 		ManualControlSettingsChannelGroupsOptions channelgroup,uint8_t *bind)
 {
-	uint32_t pios_usart_dsm_id;
+	uintptr_t pios_usart_dsm_id;
 	if (PIOS_USART_Init(&pios_usart_dsm_id, pios_usart_dsm_cfg)) {
 		PIOS_Assert(0);
 	}
 
-	uint32_t pios_dsm_id;
+	uintptr_t pios_dsm_id;
 	if (PIOS_DSM_Init(&pios_dsm_id, pios_dsm_cfg, pios_usart_com_driver,
 			pios_usart_dsm_id, *proto, *bind)) {
 		PIOS_Assert(0);
 	}
 
-	uint32_t pios_dsm_rcvr_id;
+	uintptr_t pios_dsm_rcvr_id;
 	if (PIOS_RCVR_Init(&pios_dsm_rcvr_id, &pios_dsm_rcvr_driver, pios_dsm_id)) {
 		PIOS_Assert(0);
 	}
@@ -411,7 +411,7 @@ void PIOS_Board_Init(void) {
 	usb_hid_present = true;
 #endif
 
-	uint32_t pios_usb_id;
+	uintptr_t pios_usb_id;
 	PIOS_USB_Init(&pios_usb_id, PIOS_BOARD_HW_DEFS_GetUsbCfg(bdinfo->board_rev));
 
 #if defined(PIOS_INCLUDE_USB_CDC)
@@ -431,7 +431,7 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_USB_VCPPORT_USBTELEMETRY:
 #if defined(PIOS_INCLUDE_COM)
 		{
-			uint32_t pios_usb_cdc_id;
+			uintptr_t pios_usb_cdc_id;
 			if (PIOS_USB_CDC_Init(&pios_usb_cdc_id, &pios_usb_cdc_cfg, pios_usb_id)) {
 				PIOS_Assert(0);
 			}
@@ -450,7 +450,7 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_USB_VCPPORT_COMBRIDGE:
 #if defined(PIOS_INCLUDE_COM)
 		{
-			uint32_t pios_usb_cdc_id;
+			uintptr_t pios_usb_cdc_id;
 			if (PIOS_USB_CDC_Init(&pios_usb_cdc_id, &pios_usb_cdc_cfg, pios_usb_id)) {
 				PIOS_Assert(0);
 			}
@@ -470,7 +470,7 @@ void PIOS_Board_Init(void) {
 #if defined(PIOS_INCLUDE_COM)
 #if defined(PIOS_INCLUDE_DEBUG_CONSOLE)
 		{
-			uint32_t pios_usb_cdc_id;
+			uintptr_t pios_usb_cdc_id;
 			if (PIOS_USB_CDC_Init(&pios_usb_cdc_id, &pios_usb_cdc_cfg, pios_usb_id)) {
 				PIOS_Assert(0);
 			}
@@ -505,7 +505,7 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_USB_HIDPORT_USBTELEMETRY:
 #if defined(PIOS_INCLUDE_COM)
 		{
-			uint32_t pios_usb_hid_id;
+			uintptr_t pios_usb_hid_id;
 			if (PIOS_USB_HID_Init(&pios_usb_hid_id, &pios_usb_hid_cfg, pios_usb_id)) {
 				PIOS_Assert(0);
 			}
@@ -549,15 +549,15 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_UART1_SBUS:
 #if defined(PIOS_INCLUDE_SBUS) && defined(PIOS_INCLUDE_USART)
 		{
-			uint32_t pios_usart_sbus_id;
+			uintptr_t pios_usart_sbus_id;
 			if (PIOS_USART_Init(&pios_usart_sbus_id, &pios_usart1_sbus_cfg)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_id;
+			uintptr_t pios_sbus_id;
 			if (PIOS_SBus_Init(&pios_sbus_id, &pios_usart1_sbus_aux_cfg, &pios_usart_com_driver, pios_usart_sbus_id)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_rcvr_id;
+			uintptr_t pios_sbus_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_sbus_rcvr_id, &pios_sbus_rcvr_driver, pios_sbus_id)) {
 				PIOS_Assert(0);
 			}
@@ -634,15 +634,15 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_UART2_SBUS:
 #if defined(PIOS_INCLUDE_SBUS) && defined(PIOS_INCLUDE_USART)
 		{
-			uint32_t pios_usart_sbus_id;
+			uintptr_t pios_usart_sbus_id;
 			if (PIOS_USART_Init(&pios_usart_sbus_id, &pios_usart2_sbus_cfg)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_id;
+			uintptr_t pios_sbus_id;
 			if (PIOS_SBus_Init(&pios_sbus_id, &pios_usart2_sbus_aux_cfg, &pios_usart_com_driver, pios_usart_sbus_id)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_rcvr_id;
+			uintptr_t pios_sbus_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_sbus_rcvr_id, &pios_sbus_rcvr_driver, pios_sbus_id)) {
 				PIOS_Assert(0);
 			}
@@ -718,15 +718,15 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_UART3_SBUS:
 #if defined(PIOS_INCLUDE_SBUS) && defined(PIOS_INCLUDE_USART)
 		{
-			uint32_t pios_usart_sbus_id;
+			uintptr_t pios_usart_sbus_id;
 			if (PIOS_USART_Init(&pios_usart_sbus_id, &pios_usart3_sbus_cfg)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_id;
+			uintptr_t pios_sbus_id;
 			if (PIOS_SBus_Init(&pios_sbus_id, &pios_usart3_sbus_aux_cfg, &pios_usart_com_driver, pios_usart_sbus_id)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_rcvr_id;
+			uintptr_t pios_sbus_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_sbus_rcvr_id, &pios_sbus_rcvr_driver, pios_sbus_id)) {
 				PIOS_Assert(0);
 			}
@@ -802,15 +802,15 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_UART4_SBUS:
 #if defined(PIOS_INCLUDE_SBUS) && defined(PIOS_INCLUDE_USART)
 		{
-			uint32_t pios_usart_sbus_id;
+			uintptr_t pios_usart_sbus_id;
 			if (PIOS_USART_Init(&pios_usart_sbus_id, &pios_usart4_sbus_cfg)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_id;
+			uintptr_t pios_sbus_id;
 			if (PIOS_SBus_Init(&pios_sbus_id, &pios_usart4_sbus_aux_cfg, &pios_usart_com_driver, pios_usart_sbus_id)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_rcvr_id;
+			uintptr_t pios_sbus_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_sbus_rcvr_id, &pios_sbus_rcvr_driver, pios_sbus_id)) {
 				PIOS_Assert(0);
 			}
@@ -886,15 +886,15 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_UART5_SBUS:
 #if defined(PIOS_INCLUDE_SBUS) && defined(PIOS_INCLUDE_USART)
 		{
-			uint32_t pios_usart_sbus_id;
+			uintptr_t pios_usart_sbus_id;
 			if (PIOS_USART_Init(&pios_usart_sbus_id, &pios_usart5_sbus_cfg)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_id;
+			uintptr_t pios_sbus_id;
 			if (PIOS_SBus_Init(&pios_sbus_id, &pios_usart5_sbus_aux_cfg, &pios_usart_com_driver, pios_usart_sbus_id)) {
 				PIOS_Assert(0);
 			}
-			uint32_t pios_sbus_rcvr_id;
+			uintptr_t pios_sbus_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_sbus_rcvr_id, &pios_sbus_rcvr_driver, pios_sbus_id)) {
 				PIOS_Assert(0);
 			}
@@ -962,10 +962,10 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_RCVRPORT_PWM:
 #if defined(PIOS_INCLUDE_PWM)
 		{
-			uint32_t pios_pwm_id;
+			uintptr_t pios_pwm_id;
 			PIOS_PWM_Init(&pios_pwm_id, &pios_pwm_cfg);
 
-			uint32_t pios_pwm_rcvr_id;
+			uintptr_t pios_pwm_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_pwm_rcvr_id, &pios_pwm_rcvr_driver, pios_pwm_id)) {
 				PIOS_Assert(0);
 			}
@@ -977,10 +977,10 @@ void PIOS_Board_Init(void) {
 	case HWFLYINGF3_RCVRPORT_PPMOUTPUTS:
 #if defined(PIOS_INCLUDE_PPM)
 		{
-			uint32_t pios_ppm_id;
+			uintptr_t pios_ppm_id;
 			PIOS_PPM_Init(&pios_ppm_id, &pios_ppm_cfg);
 
-			uint32_t pios_ppm_rcvr_id;
+			uintptr_t pios_ppm_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_ppm_rcvr_id, &pios_ppm_rcvr_driver, pios_ppm_id)) {
 				PIOS_Assert(0);
 			}
@@ -992,10 +992,10 @@ void PIOS_Board_Init(void) {
 		/* This is a combination of PPM and PWM inputs */
 #if defined(PIOS_INCLUDE_PPM)
 		{
-			uint32_t pios_ppm_id;
+			uintptr_t pios_ppm_id;
 			PIOS_PPM_Init(&pios_ppm_id, &pios_ppm_cfg);
 
-			uint32_t pios_ppm_rcvr_id;
+			uintptr_t pios_ppm_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_ppm_rcvr_id, &pios_ppm_rcvr_driver, pios_ppm_id)) {
 				PIOS_Assert(0);
 			}
@@ -1004,10 +1004,10 @@ void PIOS_Board_Init(void) {
 #endif	/* PIOS_INCLUDE_PPM */
 #if defined(PIOS_INCLUDE_PWM)
 		{
-			uint32_t pios_pwm_id;
+			uintptr_t pios_pwm_id;
 			PIOS_PWM_Init(&pios_pwm_id, &pios_pwm_with_ppm_cfg);
 
-			uint32_t pios_pwm_rcvr_id;
+			uintptr_t pios_pwm_rcvr_id;
 			if (PIOS_RCVR_Init(&pios_pwm_rcvr_id, &pios_pwm_rcvr_driver, pios_pwm_id)) {
 				PIOS_Assert(0);
 			}
@@ -1020,9 +1020,9 @@ void PIOS_Board_Init(void) {
 
 #if defined(PIOS_INCLUDE_GCSRCVR)
 	GCSReceiverInitialize();
-	uint32_t pios_gcsrcvr_id;
+	uintptr_t pios_gcsrcvr_id;
 	PIOS_GCSRCVR_Init(&pios_gcsrcvr_id);
-	uint32_t pios_gcsrcvr_rcvr_id;
+	uintptr_t pios_gcsrcvr_rcvr_id;
 	if (PIOS_RCVR_Init(&pios_gcsrcvr_rcvr_id, &pios_gcsrcvr_rcvr_driver, pios_gcsrcvr_id)) {
 		PIOS_Assert(0);
 	}
